+++ platform/jvm/capture/src/main/kotlin/io/bitdrift/capture/LoggerImpl.kt
@@ -170,8 +170,8 @@ internal class LoggerImpl(
                 // case of key conflicts.
                 ootbFieldProviders =
                     listOf(
-                        clientAttributes,
                         networkAttributes,
+                        FieldProvider { clientAttributes.dynamicFields() },
                     ),
                 errorHandler = errorHandler,
                 customFieldProviders = fieldProviders,
@@ -240,6 +240,9 @@ internal class LoggerImpl(
                 clientAttributes.osVersion,
                 clientAttributes.manufacturer,
                 clientAttributes.model,
+                clientAttributes.appVersionCode,
+                clientAttributes.osApiLevel,
+                clientAttributes.architecture,
                 network,
                 preferences,
                 localErrorReporter,