+++ platform/swift/source/Logger.swift
@@ -129,22 +129,14 @@ public final class Logger {
         self.timeProvider = timeProvider
         let start = timeProvider.uptime()
 
-        // Order of providers matters in here, the latter in the list the higher their priority in
-        // case of key conflicts.
         let appStateAttributes = AppStateAttributes()
         let clientAttributes = ClientAttributes()
         let deviceAttributes = DeviceAttributes()
         let networkAttributes = NetworkAttributes()
-        let ootbFieldProviders: [FieldProvider] = [
-            appStateAttributes,
-            clientAttributes,
-            deviceAttributes,
-            networkAttributes,
-        ]
 
         let metadataProvider = MetadataProviderController(
             dateProvider: dateProvider ?? SystemDateProvider(),
-            ootbFieldProviders: ootbFieldProviders,
+            ootbFieldProviders: [appStateAttributes, deviceAttributes, networkAttributes],
             customFieldProviders: fieldProviders
         )
 
@@ -194,6 +186,7 @@ public final class Logger {
             eventsListenerTarget: self.eventsListenerTarget,
             appID: clientAttributes.appID,
             releaseVersion: clientAttributes.appVersion,
+            buildNumber: clientAttributes.buildNumber,
             osVersion: clientAttributes.osVersion,
             model: deviceAttributes.hardwareVersion,
             network: network,