+++ platform/swift/source/default/attributes/DeviceAttributes.swift
@@ -42,8 +42,6 @@ final class DeviceAttributes {
 extension DeviceAttributes: FieldProvider {
     public func getFields() -> Fields {
         return [
-            /// The iPhone device model (e.g. iPhone13,1)
-            "model": self.hardwareVersion,
             /// The device locale (e.g. en_US)
             "_locale": self.locale.load(),
         ]