+++ test/benchmark/src/bin/live_benchmark.rs
@@ -43,6 +43,7 @@ fn test_live_match_performance(c: &mut Criterion) {
     session_strategy: Arc::new(Strategy::fixed(".", Arc::new(UUIDCallbacks))),
     store,
     metadata_provider,
+    initial_ootb_fields: [].into(),
     resource_utilization_target: Box::new(bd_test_helpers::resource_utilization::EmptyTarget),
     session_replay_target: Box::new(bd_test_helpers::session_replay::NoOpTarget),
     events_listener_target: Box::new(bd_test_helpers::events::NoOpListenerTarget),