+++ test/benchmark/src/bin/logger_benchmark.rs
@@ -57,6 +57,7 @@ fn simple_log(c: &mut Criterion) {
       timestamp: time::OffsetDateTime::now_utc().into(),
       ..Default::default()
     }),
+    initial_ootb_fields: [].into(),
     resource_utilization_target: Box::new(bd_test_helpers::resource_utilization::EmptyTarget),
     session_replay_target: Box::new(bd_test_helpers::session_replay::NoOpTarget),
     events_listener_target: Box::new(bd_test_helpers::events::NoOpListenerTarget),
@@ -99,6 +100,7 @@ fn with_matcher_and_buffer(c: &mut Criterion) {
       timestamp: time::OffsetDateTime::now_utc().into(),
       ..Default::default()
     }),
+    initial_ootb_fields: [].into(),
     resource_utilization_target: Box::new(bd_test_helpers::resource_utilization::EmptyTarget),
     session_replay_target: Box::new(bd_test_helpers::session_replay::NoOpTarget),
     events_listener_target: Box::new(bd_test_helpers::events::NoOpListenerTarget),